Clean up the database interface; use a custom row factory.
We know the schema of the main table being used, so there's no reason
why can't just return a dictionary based on that instead. I'm unsure
whether there are any speed benefits with this as opposed to
sqlite3.Row, but the conversion into JSON would seem to be considerably
easier using this factory.