Alphagram function.
An alphagram is an alphabetically sorted 'anagram' of a word. Thus, CAT
becomes ACT, DOG becomes DGO, and GOODBYE becomes BDEGOOY. For simple
anagram searches, it's enough to convert to an alaphagram and compare
against the alphagram column of the database.

Simple anagram searches consist of those that do not include wildcards.