Define different handing of Zyzzyva's anagram match patterns.
They're unordered regular expressions, which means that regular
expressions are rubbish for actually matching anagrams. However, there's
no reason why it should be a *complicated* implementation.

'PatternMatch' simply translates values into an SQLite pattern, which is
perfect.

'AnagramMatch' (and its evil cousin, 'SubanagramMatch') require more
complex parsing. To be honest, I don't really totally understand how the
core of libzyzzyva handles either; from what I gather, it's a form of
acyclic word graph/trie system, and it somehow generates these graphs
based on the lexicon. Whether or not these are pre-generated or are
contained within the database, I'm uncertain.

However, the patterns can be broken down into the following components,
each of which is optional:

 - Number and specific letters ("ZAE" thus has three letters, which are
   "Z", "A", and "E")
 - Number of "blanks" ("ZAE?" thus has one blank)
 - Whether or not a wildcard is contained ("ZAE*" has one wild-card;
   "*ZAE*" is identical to "ZAE*")
 - Number and content of contained character sets ("Z[AEIOU][AEIOU]",
   thus contains two sets, whose content is identical (but duplicated)
   with "A", "E", "I", "O", "U").

From these components it is possible to determine the "length" of the
word this pattern will accept; "Z[AEIOU][AEIOU]?" will accept any 4
letter word -- and unless it is defined as a subanagram, no less than 4
letter words; "Z[AEIOU][AEIOU]?*" will accept any word of four
characters or more.

Thus, once we have a pattern, we also have a soft limit for our database
search (an even harder one if the query maker is asking for a specific
length as well), which reduces the number of results to check.

Using sqlite's create_function interface, we can then generate a
checker function that returns a boolean True or False value for each
checked worked.

Finally, this allows us to limit results based on a pattern!

The Pattern (and SubPattern) objects are themselves stateful per-letter,
but not per-word. However, as the object is 'reset' after every word (by
not actually changing the object but instead cloning it, per-word), the
same Pattern object can be used for every single word being tested.

This should hopefully decrease the overhead of re-parsing the pattern
for every word. However, it's also possible that we actually need to
re-set the object instead of cloning: it depends where the slow-downs
are, etc.