Include the fileinput plugin to power the load button.
It's slightly modified, removing the span, etc, code, to rely instead
on the jQuery ui-provided buttons (possibly this code predates the
addition of buttons to the ui-core).

This now moves the buttons into a table ("ew, tables!") and includes a
little bit of CSS-hackery and some tweaking of sizes to make everything
line up properly.

It probably now ceases to work on other operating systems and browsers.
Time to test!